# Conflicts:
#	src/main/java/com/gokkan/gokkan/domain/expertInfo/domain/ExpertInfo.java
#	src/main/java/com/gokkan/gokkan/domain/expertInfo/domain/dto/ExpertInfoDto.java
#	src/main/java/com/gokkan/gokkan/domain/expertInfo/service/ExpertInfoService.java
#	src/main/java/com/gokkan/gokkan/domain/member/domain/Member.java
#	src/main/java/com/gokkan/gokkan/domain/style/service/ExpertStyleService.java
#	src/test/java/com/gokkan/gokkan/domain/style/service/ExpertStyleServiceTest.java

+++ src/main/java/com/gokkan/gokkan/domain/item/service/ItemService.java
@@ -5,6 +5,8 @@ import static com.gokkan.gokkan.domain.item.dto.ItemDto.Response;
 import com.gokkan.gokkan.domain.category.domain.Category;
 import com.gokkan.gokkan.domain.image.domain.ImageCheck;
 import com.gokkan.gokkan.domain.image.domain.ImageItem;
+import com.gokkan.gokkan.domain.image.repository.ImageCheckRepository;
+import com.gokkan.gokkan.domain.image.repository.ImageItemRepository;
 import com.gokkan.gokkan.domain.image.service.ImageCheckService;
 import com.gokkan.gokkan.domain.image.service.ImageItemService;
 import com.gokkan.gokkan.domain.item.domain.Item;
@@ -13,6 +15,7 @@ import com.gokkan.gokkan.domain.item.dto.ItemDto.UpdateRequest;
 import com.gokkan.gokkan.domain.item.exception.ItemErrorCode;
 import com.gokkan.gokkan.domain.item.repository.ItemRepository;
 import com.gokkan.gokkan.domain.style.domain.StyleItem;
+import com.gokkan.gokkan.domain.style.repository.StyleItemRepository;
 import com.gokkan.gokkan.global.exception.exception.RestApiException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -24,6 +27,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class ItemService {
 
 	private final ItemRepository itemRepository;
+	private final StyleItemRepository styleItemRepository;
+	private final ImageItemRepository imageItemRepository;
+	private final ImageCheckRepository imageCheckRepository;
 	private final ImageItemService imageItemService;
 	private final ImageCheckService imageCheckService;
 
@@ -31,17 +37,12 @@ public class ItemService {
 	@Transactional
 	public Response create(
 		CreateRequest request,
+		List<ImageItem> imageItems,
+		List<ImageCheck> imageChecks,
 		Category category,
-		List<StyleItem> styleItems,
-		List<ImageItem> imageItemUrls,
-		List<ImageCheck> imageCheckUrls) {
-
-		Item item = request.toItem();
-		item.setCategory(category);
-		item.addImageItems(imageItemUrls);
-		item.addImageChecks(imageCheckUrls);
-		item.addStyleItem(styleItems);
-
+		List<StyleItem> styleItems) {
+		Item item = itemRepository.save(request.toItem(category));
+		saveItemRelations(imageItems, imageChecks, styleItems, item);
 		return Response.toResponse(itemRepository.save(item));
 	}
 
@@ -67,26 +68,34 @@ public class ItemService {
 	@Transactional
 	public Response update(
 		UpdateRequest request,
-		Category category,
-		List<StyleItem> styleItems,
 		List<ImageItem> imageItems,
-		List<ImageCheck> imageChecks) {
+		List<ImageCheck> imageChecks,
+		Category category,
+		List<StyleItem> styleItems) {
 
 		Item item = getItem(request.getItemId());
-		item = request.toItem(item);
-		item.setCategory(category);
+		item = request.toItem(item, category);
 
-		for (ImageItem imageItem : item.getImageItems()) {
-			imageItemService.delete(imageItem);
+		List<StyleItem> styleItemsSaved = item.getStyleItems();
+		if (styleItemsSaved != null && styleItemsSaved.size() != 0) {
+			styleItemRepository.deleteAll(styleItemsSaved);
 		}
 
-		for (ImageCheck imageCheck : item.getImageChecks()) {
-			imageCheckService.delete(imageCheck);
+		List<ImageItem> imageItemsSaved = item.getImageItems();
+		if (imageItemsSaved != null && imageItemsSaved.size() != 0) {
+			for (ImageItem imageItem : imageItemsSaved) {
+				imageItemService.delete(imageItem);
+			}
 		}
 
-		item.addImageItems(imageItems);
-		item.addImageChecks(imageChecks);
-		item.addStyleItem(styleItems);
+		List<ImageCheck> imageChecksSaved = item.getImageChecks();
+		if (imageChecksSaved != null && imageChecksSaved.size() != 0) {
+			for (ImageCheck imageCheck : imageChecksSaved) {
+				imageCheckService.delete(imageCheck);
+			}
+		}
+
+		saveItemRelations(imageItems, imageChecks, styleItems, item);
 
 		return Response.toResponse(itemRepository.save(item));
 	}
@@ -94,5 +103,20 @@ public class ItemService {
 	private Item getItem(Long itemId) {
 		return itemRepository.findById(itemId)
 			.orElseThrow((() -> new RestApiException(ItemErrorCode.NOT_FOUND_ITEM)));
+	}
+
+	private void saveItemRelations(
+		List<ImageItem> imageItems,
+		List<ImageCheck> imageChecks,
+		List<StyleItem> styleItems,
+		Item item) {
+
+		item.addStyleItem(styleItems);
+		item.addImageItems(imageItems);
+		item.addImageChecks(imageChecks);
+
+		imageItemRepository.saveAll(imageItems);
+		imageCheckRepository.saveAll(imageChecks);
+		styleItemRepository.saveAll(styleItems);
 	}
 }
